+++ config/event/config.go
@@ -1,9 +1,7 @@
 package event
 
 import (
-	"context"
-	"fmt"
-
+	c "github.com/crossplane-contrib/provider-pagerduty/config/common"
 	"github.com/crossplane/upjet/pkg/config"
 )
 
@@ -52,6 +50,7 @@ func Configure(p *config.Provider) {
 	p.AddResourceConfigurator("pagerduty_event_orchestration", func(r *config.Resource) {
 
 		r.ShortGroup = ShortGroup
+		r.ExternalName = c.SplitExternalNameFromId()
 		r.References = config.References{
 			"team": {
 				Type: "github.com/crossplane-contrib/provider-pagerduty/apis/team/v1alpha1.Team",
@@ -62,9 +61,7 @@ func Configure(p *config.Provider) {
 	p.AddResourceConfigurator("pagerduty_event_orchestration_global_cache_variable", func(r *config.Resource) {
 
 		r.ShortGroup = ShortGroup
-		r.ExternalName.GetIDFn = func(_ context.Context, externalName string, parameters map[string]any, _ map[string]any) (string, error) {
-			return fmt.Sprintf("%s:%s", parameters["event_orchestration"].(string), externalName), nil
-		}
+		r.ExternalName = c.SplitExternalNameFromId()
 		r.References = config.References{
 			"event_orchestration": {
 				Type: "Orchestration",
@@ -75,9 +72,7 @@ func Configure(p *config.Provider) {
 	p.AddResourceConfigurator("pagerduty_event_orchestration_global", func(r *config.Resource) {
 
 		r.ShortGroup = ShortGroup
-		r.ExternalName.GetIDFn = func(_ context.Context, externalName string, parameters map[string]any, _ map[string]any) (string, error) {
-			return fmt.Sprintf("%s:%s", parameters["event_orchestration"].(string), externalName), nil
-		}
+		r.ExternalName = c.SplitExternalNameFromId()
 		r.References = config.References{
 			"event_orchestration": {
 				Type: "Orchestration",
@@ -88,9 +83,7 @@ func Configure(p *config.Provider) {
 	p.AddResourceConfigurator("pagerduty_event_orchestration_integration", func(r *config.Resource) {
 
 		r.ShortGroup = ShortGroup
-		r.ExternalName.GetIDFn = func(_ context.Context, externalName string, parameters map[string]any, _ map[string]any) (string, error) {
-			return fmt.Sprintf("%s:%s", parameters["event_orchestration"].(string), externalName), nil
-		}
+		r.ExternalName = c.SplitExternalNameFromId()
 		r.References = config.References{
 			"event_orchestration": {
 				Type: "Orchestration",
@@ -101,9 +94,7 @@ func Configure(p *config.Provider) {
 	p.AddResourceConfigurator("pagerduty_event_orchestration_service_cache_variable", func(r *config.Resource) {
 
 		r.ShortGroup = ShortGroup
-		r.ExternalName.GetIDFn = func(_ context.Context, externalName string, parameters map[string]any, _ map[string]any) (string, error) {
-			return fmt.Sprintf("%s:%s", parameters["service"].(string), externalName), nil
-		}
+		r.ExternalName = c.SplitExternalNameFromId()
 		r.References = config.References{
 			"service": {
 				Type: "OrchestrationService",
